requirements.txt
altgraph==0.16.1
certifi==2018.8.24
chardet==3.0.4
future==0.16.0
idna==2.7
macholib==1.11
pefile==2018.8.8
PyInstaller==3.3.1
pyodbc==4.0.24
pypiwin32==223
python-gitlab==1.6.0
pywin32==223
requests==2.19.1
six==1.11.0
urllib3==1.23
